requirements.txt
pre-commit
mkdocs
mkdocs-material
proselint
Pillow
